Move load and save out of model to identify necessary changes
NOTE: This code will not compile!

I decided it would be a good idea to move the code out
so that we can better identify what dependencies
load and save require from the model.

I also exposed the connections since we need them
exposed for saving and we will want them for
build mode if we wish to add deletion of connections.