Fix placement checking on moves
checkPlacement() will check the position of the provided ball or gizmo
against the already existing ones. We must remove the ball or gizmo we
are checking from the model's internal sets, check the placement, and
then add it back.

Alternatively, we could just check that the board is in a valid state at
any given moment. This would require to alter the board by adding the
ball or gizmo before checking if the addition is valid. On invalid
additions, this would leave the model in an invalid state.

And invalid state sucks.